Extract bootstrap methods to a superclass and let those tests extends superclass
