+++ hadoop-ozone/ozone-manager/src/test/java/org/apache/hadoop/ozone/om/response/volume/TestOMVolumeSetOwnerResponse.java
@@ -18,10 +18,6 @@
 
 package org.apache.hadoop.ozone.om.response.volume;
 
-import org.apache.hadoop.hdds.conf.OzoneConfiguration;
-import org.apache.hadoop.ozone.om.OMConfigKeys;
-import org.apache.hadoop.ozone.om.OMMetadataManager;
-import org.apache.hadoop.ozone.om.OmMetadataManagerImpl;
 import org.apache.hadoop.ozone.om.helpers.OmVolumeArgs;
 import org.apache.hadoop.ozone.protocol.proto.OzoneManagerProtocolProtos;
 import org.apache.hadoop.ozone.protocol.proto.OzoneManagerProtocolProtos
@@ -30,14 +26,9 @@ import org.apache.hadoop.ozone.protocol.proto.OzoneManagerProtocolProtos
     .OMResponse;
 import org.apache.hadoop.ozone.storage.proto.OzoneManagerStorageProtos.PersistedUserVolumeInfo;
 import org.apache.hadoop.util.Time;
-import org.apache.hadoop.hdds.utils.db.BatchOperation;
 import org.apache.hadoop.hdds.utils.db.Table;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
 
-import java.nio.file.Path;
 import java.util.UUID;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -45,30 +36,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 /**
  * This class tests OMVolumeCreateResponse.
  */
-public class TestOMVolumeSetOwnerResponse {
-
-  @TempDir
-  private Path folder;
-
-  private OMMetadataManager omMetadataManager;
-  private BatchOperation batchOperation;
-
-  @BeforeEach
-  public void setup() throws Exception {
-    OzoneConfiguration ozoneConfiguration = new OzoneConfiguration();
-    ozoneConfiguration.set(OMConfigKeys.OZONE_OM_DB_DIRS,
-        folder.toAbsolutePath().toString());
-    omMetadataManager = new OmMetadataManagerImpl(ozoneConfiguration, null);
-    batchOperation = omMetadataManager.getStore().initBatchOperation();
-  }
-
-  @AfterEach
-  public void tearDown() {
-    if (batchOperation != null) {
-      batchOperation.close();
-    }
-  }
-
+public class TestOMVolumeSetOwnerResponse extends TestOMVolumeResponse {
 
   @Test
   public void testAddToDBBatch() throws Exception {
@@ -94,25 +62,24 @@ public class TestOMVolumeSetOwnerResponse {
         new OMVolumeCreateResponse(omResponse, omVolumeArgs, volumeList);
 
 
-
     String newOwner = "user2";
     PersistedUserVolumeInfo newOwnerVolumeList =
         PersistedUserVolumeInfo.newBuilder()
-        .setObjectID(1)
-        .setUpdateID(1)
-        .addVolumeNames(volumeName).build();
+            .setObjectID(1)
+            .setUpdateID(1)
+            .addVolumeNames(volumeName).build();
     PersistedUserVolumeInfo oldOwnerVolumeList =
         PersistedUserVolumeInfo.newBuilder()
-        .setObjectID(2)
-        .setUpdateID(2)
-        .build();
+            .setObjectID(2)
+            .setUpdateID(2)
+            .build();
     OmVolumeArgs newOwnerVolumeArgs = OmVolumeArgs.newBuilder()
         .setOwnerName(newOwner).setAdminName(newOwner)
         .setVolume(volumeName).setCreationTime(omVolumeArgs.getCreationTime())
         .build();
 
     OMVolumeSetOwnerResponse omVolumeSetOwnerResponse =
-        new OMVolumeSetOwnerResponse(omResponse, oldOwner,  oldOwnerVolumeList,
+        new OMVolumeSetOwnerResponse(omResponse, oldOwner, oldOwnerVolumeList,
             newOwnerVolumeList, newOwnerVolumeArgs);
 
     omVolumeCreateResponse.addToDBBatch(omMetadataManager, batchOperation);
@@ -155,6 +122,4 @@ public class TestOMVolumeSetOwnerResponse {
     assertEquals(0, omMetadataManager.countRowsInTable(
         omMetadataManager.getVolumeTable()));
   }
-
-
 }
